--- chatbot_attention_mechanism/chabot.py
@@ -1,231 +1,0 @@
-import torch
-import torch.nn as nn
-from torch.nn import functional as F
-import pandas as pd
-
-batch_size = 4
-block_size = 16
-max_iters = 5000
-eval_interval = 100
-learning_rate = 1e-3
-device = 'cuda' if torch.cuda.is_available() else 'cpu'
-eval_iters = 200
-n_embd = 64
-n_head = 4
-n_layer = 4
-dropout = 0.0
-
-nRowsRead = None
-text1 = pd.read_csv(
-    '/content/drive/MyDrive/Colab Notebooks/Mental_Health_FAQ.csv',
-    delimiter=',', nrows=nRowsRead
-)
-text1.dataframeName = 'Mental_Health_FAQ.csv'
-nRow, nCol = text1.shape
-
-nRowsRead = None
-text2 = pd.read_csv(
-    '/content/drive/MyDrive/Colab Notebooks/medquad.csv',
-    delimiter=',', nrows=nRowsRead
-)
-text2.dataframeName = 'medquad.csv'
-
-text1 = text1.drop('Question_ID', axis=1)
-text2 = text2.drop(['source', 'focus_area'], axis=1)
-text2 = text2.rename(columns={"question": "Questions", "answer": "Answers"})
-
-text = pd.concat([text1, text2], axis=0)
-text = ''.join(text['Questions'].astype(str) + text['Answers'].astype(str))
-
-torch.manual_seed(1337)
-chars = sorted(list(set(text)))
-vocab_size = len(chars)
-
-stoi = {ch: i for i, ch in enumerate(chars)}
-itos = {i: ch for i in enumerate(chars)}
-
-
-def encode(s):
-    return [stoi[c] for c in s]
-
-
-def decode(lst):
-    return ''.join([itos[i] for i in lst])
-
-
-data = torch.tensor(encode(text), dtype=torch.long)
-n = int(0.9 * len(data))
-train_data = data[:n]
-val_data = data[n:]
-
-
-def get_batch(split):
-    data = train_data if split == 'train' else val_data
-    ix = torch.randint(len(data) - block_size, (batch_size,))
-    x = torch.stack([data[i:i + block_size] for i in ix])
-    y = torch.stack([data[i + 1:i + block_size + 1] for i in ix])
-    return x, y
-
-
-@torch.no_grad()
-def estimate_loss():
-    out = {}
-    model.eval()
-    for split in ['train', 'val']:
-        losses = torch.zeros(eval_iters)
-        for k in range(eval_iters):
-            X, Y = get_batch(split)
-            logits, loss = model(X, Y)
-            losses[k] = loss.item()
-        out[split] = losses.mean()
-    model.train()
-    return out
-
-
-class Head(nn.Module):
-    def __init__(self, head_size):
-        super().__init__()
-        self.key = nn.Linear(n_embd, head_size, bias=False)
-        self.query = nn.Linear(n_embd, head_size, bias=False)
-        self.value = nn.Linear(n_embd, head_size, bias=False)
-        self.register_buffer(
-            'tril', torch.tril(torch.ones(block_size, block_size))
-        )
-        self.dropout = nn.Dropout(dropout)
-
-    def forward(self, x):
-        B, T, C = x.shape
-        k = self.key(x)
-        q = self.query(x)
-        wei = q @ k.transpose(-2, -1) * C ** -0.5
-        wei = wei.masked_fill(self.tril[:T, :T] == 0, float('-inf'))
-        wei = F.softmax(wei, dim=-1)
-        wei = self.dropout(wei)
-        v = self.value(x)
-        out = wei @ v
-        return out
-
-
-class MultiHeadAttention(nn.Module):
-    def __init__(self, num_heads, head_size):
-        super().__init__()
-        self.heads = nn.ModuleList([Head(head_size) for _ in range(num_heads)])
-        self.proj = nn.Linear(n_embd, n_embd)
-        self.dropout = nn.Dropout(dropout)
-
-    def forward(self, x):
-        out = torch.cat([h(x) for h in self.heads], dim=-1)
-        out = self.dropout(self.proj(out))
-        return out
-
-
-class FeedFoward(nn.Module):
-    def __init__(self, n_embd):
-        super().__init__()
-        self.net = nn.Sequential(
-            nn.Linear(n_embd, 4 * n_embd),
-            nn.ReLU(),
-            nn.Linear(4 * n_embd, n_embd),
-            nn.Dropout(dropout),
-        )
-
-    def forward(self, x):
-        return self.net(x)
-
-
-class Block(nn.Module):
-    def __init__(self, n_embd, n_head):
-        super().__init__()
-        head_size = n_embd // n_head
-        self.sa1 = MultiHeadAttention(n_head, head_size)
-        self.sa2 = MultiHeadAttention(n_head, head_size)
-        self.ffwd = FeedFoward(n_embd)
-        self.ln1 = nn.LayerNorm(n_embd)
-        self.ln2 = nn.LayerNorm(n_embd)
-        self.ln3 = nn.LayerNorm(n_embd)
-
-    def forward(self, x):
-        x1 = self.sa1(self.ln1(x))
-        x = x + x1
-        x2 = self.sa2(self.ln3(x))
-        x = x + x2
-        x = x + self.ffwd(self.ln2(x))
-        return x
-
-
-class BigramLanguageModel(nn.Module):
-    def __init__(self):
-        super().__init__()
-        self.token_embedding_table = nn.Embedding(vocab_size, n_embd)
-        self.position_embedding_table = nn.Embedding(block_size, n_embd)
-        self.blocks = nn.Sequential(
-            *[Block(n_embd, n_head=n_head) for _ in range(n_layer)]
-        )
-        self.ln_f = nn.LayerNorm(n_embd)
-        self.lm_head = nn.Linear(n_embd, vocab_size)
-
-    def forward(self, idx, targets=None):
-        B, T = idx.shape
-        tok_emb = self.token_embedding_table(idx)
-        pos_emb = self.position_embedding_table(torch.arange(T, device=device))
-        x = tok_emb + pos_emb
-        x = self.blocks(x)
-        x = self.ln_f(x)
-        logits = self.lm_head(x)
-        if targets is None:
-            loss = None
-        else:
-            B, T, C = logits.shape
-            logits = logits.view(B * T, C)
-            targets = targets.view(B * T)
-            loss = F.cross_entropy(logits, targets)
-        return logits, loss
-
-    def generate(self, idx, max_new_tokens):
-        for _ in range(max_new_tokens):
-            idx_cond = idx[:, -block_size:]
-            logits, loss = self(idx_cond)
-            logits = logits[:, -1, :]
-            probs = F.softmax(logits, dim=-1)
-            idx_next = torch.multinomial(probs, num_samples=1)
-            idx = torch.cat((idx, idx_next), dim=1)
-        return idx
-
-
-model = BigramLanguageModel()
-m = model.to(device)
-
-print(sum(p.numel() for p in m.parameters()) / 1e6, 'M parameters')
-
-optimizer = torch.optim.AdamW(model.parameters(), lr=learning_rate)
-
-for iter in range(max_iters):
-    if iter % eval_interval == 0 or iter == max_iters - 1:
-        losses = estimate_loss()
-        print(f"step {iter}: train loss {losses['train']:.4f}, val loss "
-              f"{losses['val']:.4f}")
-
-    xb, yb = get_batch('train')
-    logits, loss = model(xb, yb)
-    optimizer.zero_grad(set_to_none=True)
-    loss.backward()
-    optimizer.step()
-
-context = torch.zeros((1, 1), dtype=torch.long, device=device)
-print(decode(m.generate(context, max_new_tokens=2000)[0].tolist()))
-
-while True:
-    question = input("Enter your question (or type 'quit' to exit): ")
-    if question.lower() == 'quit':
-        break
-
-    encoded_question = torch.tensor(
-        encode(question), dtype=torch.long, device=device
-    ).unsqueeze(0)
-
-    generated_response = m.generate(encoded_question,
-     max_new_tokens=1000)[0].tolist()
-
-    decoded_response = decode(generated_response)
-
-    print("Generated response:", decoded_response)
