+++ chatbot_attention_mechanism/chatbot/training/train_model.py
@@ -1,0 +1,40 @@
+import torch
+
+
+def get_batch(split, data, block_size, batch_size):
+    ix = torch.randint(len(data) - block_size, (batch_size,))
+    x = torch.stack([data[i:i + block_size] for i in ix])
+    y = torch.stack([data[i + 1:i + block_size + 1] for i in ix])
+    return x, y
+
+
+@torch.no_grad()
+def estimate_loss(model, train_data, val_data, eval_iters, block_size, batch_size):
+    out = {}
+    model.eval()
+    for split, data in [('train', train_data), ('val', val_data)]:
+        losses = torch.zeros(eval_iters)
+        for k in range(eval_iters):
+            X, Y = get_batch(split, data, block_size, batch_size)
+            _, loss = model(X, Y)
+            losses[k] = loss.item()
+        out[split] = losses.mean()
+    model.train()
+    return out
+
+
+def train_model(model, train_data, val_data, optimizer, max_iters,
+                eval_interval, block_size, batch_size, eval_iters):
+    for iter in range(max_iters):
+        if iter % eval_interval == 0 or iter == max_iters - 1:
+            losses = estimate_loss(
+                model, train_data, val_data, eval_iters, block_size, batch_size
+            )
+            print(f"step {iter}: train loss {losses['train']:.4f}, "
+                  f"val loss {losses['val']:.4f}")
+
+        xb, yb = get_batch('train', train_data, block_size, batch_size)
+        logits, loss = model(xb, yb)
+        optimizer.zero_grad(set_to_none=True)
+        loss.backward()
+        optimizer.step()
